refactor: 구글 자동로그인으로 인한 페이지 이동 방지 및 인증 로딩 문구 추가
